Added types for SIPBLLMs, corrected tag model name, and more.
