+++ src/llm/eventToTags.ts
@@ -2,13 +2,17 @@ import { Event } from "@prisma/client";
 import dedent from "dedent";
 import { ChatCompletionFunctions } from "openai";
 
-import { createChatCompletionWithRetry, removeArtifacts, removeBase64 } from "./utils.js";
-import { CHEAP_MODEL, MODEL } from "./utils.js";
+import { createChatCompletionWithRetry, removeArtifacts } from "./utils.js";
+import { CHEAP_MODEL } from "./utils.js";
+import { doSIPBLLMsCompletion } from "./SIPBLLMsUtils.js";
 
-export const CURRENT_MODEL_NAME = "TAG-20241002";
+export const CURRENT_MODEL_NAME = "Mixtral-GPT1.0"
+interface SIPBLLMContentTag {
+    content_tag_1?: string;
+    content_tag_2?: string;
+    justification?: string;
+}
 
-
-// PLAYsentation (Class Presentation), Logar(Concert), Alien Conspiracy
 
 const PROMPT_INTRO = dedent`
   You are a campus event tagger. Your job is to reason whether given tags apply to a specific event.
@@ -176,66 +180,6 @@ const EVENT_AMENITIES_TAG_FUNCTION: ChatCompletionFunctions = {
   }
 };
 
-function extractAndSanitizeJsonContent(content: string): string {
-  const jsonMatch = content.match(/{.*}/s);
-  if (!jsonMatch) {
-    return '';
-  }
-
-  return jsonMatch[0].replace(/\\(?!["\\/bfnrtu])/g, '\\\\');
-};
-
-export async function doCompletion(prompt: string, grammar: string): Promise<any> {
-  try {
-     const response = await fetch(`${process.env.SIPB_LLMS_API_ENDPOINT}`, {
-        method: "POST",
-        headers: {
-           "Authorization": `Bearer ${process.env.SIPB_LLMS_API_TOKEN}`,
-           "Content-Type": `application/json`,
-        },
-        body: JSON.stringify({
-           "messages": [
-              {"role": "user", "content": prompt},
-           ],
-           "stream": false,
-           "tokenize": true,
-           "stop": ["</s>", "### User Message", "### Assistant", "### Prompt"],
-           "cache_prompt": false,
-           "frequency_penalty": 0,
-           "grammar": grammar,
-           "image_data": [],
-           //"model": "mixtral",
-           "min_p": 0.05,
-           "mirostat": 0,
-           "mirostat_eta": 0.1,
-           "mirostat_tau": 5,
-           "n_predict": 1000,
-           "n_probs": 0,
-           "presence_penalty": 0,
-           "repeat_last_n": 256,
-           "repeat_penalty": 1.18,
-           "seed": -1,
-           "slot_id": -1,
-           "temperature": 0.7,
-           "tfs_z": 1,
-           "top_k": 40,
-           "top_p": 0.95,
-           "typical_p": 1,
-        }),
-     });
-
-     if (!response.ok)
-        throw new Error(`HTTP error: ${response.status}. Response: ${await response.text()}`);
-
-     const data = await response.json();
-     const content = data["choices"][0]["message"]["content"];
-     return JSON.parse(extractAndSanitizeJsonContent(content));
-
-  } catch (error) {
-     console.error(`Error with completion:`, error);
-     throw error;
-  }
-}
 
 export async function addTagsToEvent(event: Event): Promise<string[]> {
   const text = removeArtifacts(event.text);
@@ -284,19 +228,19 @@ export async function addTagsToEvent(event: Event): Promise<string[]> {
     return results;
   }
 
-  async function extractTags(prompt: string, grammar: string, allowed: string[]): Promise<string[]> {
+  async function extractContentTags(prompt: string, grammar: string, allowed: string[]): Promise<string[]> {
     const systemPrompt = prompt.replace("{INSERT TITLE HERE}", event.title);
     try {
-      const response = await doCompletion(
+      const response = await doSIPBLLMsCompletion(
         `${systemPrompt}\n\`\`\`\n${text}\n\`\`\`\n\n---------------- Response --------------\n`,
         grammar
-      )
+      ) as SIPBLLMContentTag;
 
-      const tags = [response].flatMap(tags => Object.values(tags) as string[])
-                               .filter((tag) => allowed.includes(tag));
+      const tags: Array<string> = [response.content_tag_1, response.content_tag_2]
+      .filter((tag): tag is string => tag !== undefined && allowed.includes(tag));
 
       if (process.env.DEBUG_MODE) {
-        console.log("----------Extracted Tags----------");
+        console.log("----------Extracted Content Tags----------");
         console.log(tags);
         console.log("----------Justification---------");
         console.log(response["justification"]);
@@ -313,7 +257,7 @@ export async function addTagsToEvent(event: Event): Promise<string[]> {
   // TODO: Update all types of tags to use SIPB LLMs endpoint in doCompletion by invoking extractTags
   const [formTags, contentTags, amenitiesTags] = await Promise.all([
     twoStagePrompt(FORM_TAG_PROMPT, EVENT_FORM_TAG_FUNCTION, ACCEPTABLE_FORM_TAGS),
-    extractTags(CONTENT_TAG_PROMPT, EVENT_CONTENT_TAG_GRAMMAR, ACCEPTABLE_CONTENT_TAGS),
+    extractContentTags(CONTENT_TAG_PROMPT, EVENT_CONTENT_TAG_GRAMMAR, ACCEPTABLE_CONTENT_TAGS),
     twoStagePrompt(AMENITIES_TAG_PROMPT, EVENT_AMENITIES_TAG_FUNCTION, ACCEPTABLE_AMENITIES_TAGS)
   ]);
